refactor GM script names remove GMScriptCallback class (unused) move gm scipts lua API to GMActions
