+++ EmptyEpsilon.cbp
@@ -450,8 +450,6 @@
 		<Unit filename="EmptyEpsilon.rc">
 			<Option compilerVar="WINDRES" />
 		</Unit>
-		<Unit filename="src/GMScriptCallback.cpp" />
-		<Unit filename="src/GMScriptCallback.h" />
 		<Unit filename="src/ai/ai.cpp" />
 		<Unit filename="src/ai/ai.h" />
 		<Unit filename="src/ai/aiFactory.cpp" />
